+++ geode-wan/src/distributedTest/java/org/apache/geode/cache/CacheXml80GatewayDUnitTest.java
@@ -31,8 +31,8 @@ import org.apache.geode.cache.wan.GatewayReceiverFactory;
 import org.apache.geode.cache.wan.GatewaySender;
 import org.apache.geode.cache.wan.GatewaySenderFactory;
 import org.apache.geode.cache.wan.GatewayTransportFilter;
-import org.apache.geode.cache30.CacheXml70DUnitTest;
 import org.apache.geode.cache30.CacheXmlTestCase;
+import org.apache.geode.cache30.CacheXml70DUnitTestHelper;
 import org.apache.geode.cache30.MyGatewayTransportFilter1;
 import org.apache.geode.cache30.MyGatewayTransportFilter2;
 import org.apache.geode.internal.cache.xmlcache.CacheCreation;
@@ -85,7 +85,7 @@ public class CacheXml80GatewayDUnitTest extends CacheXmlTestCase {
     String id = getName();
     AsyncEventQueueFactory factory = cache.createAsyncEventQueueFactory();
     factory.setGatewayEventSubstitutionListener(new MyGatewayEventSubstitutionFilter());
-    AsyncEventQueue queue = factory.create(id, new CacheXml70DUnitTest.MyAsyncEventListener());
+    AsyncEventQueue queue = factory.create(id, new CacheXml70DUnitTestHelper.MyAsyncEventListener());
 
     // Verify the GatewayEventSubstitutionFilter is set on the AsyncEventQueue.
     assertNotNull(queue.getGatewayEventSubstitutionFilter());
